Update delay.m
The routine now also output the minimum of intercorrelation. When two waveforms are similar but very much out of phase, the minimum of intercorrelation will be larger the maximum in absolute value. It can be useful information to have when interpreting results.

The routine is also modified to make sure the function does not bug when the maximum or minimum are found on either edge of the signal (first or last sample). This can happen if you have a (very) noisy signal typically.